+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoatController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoatacController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateCoverController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateParticipantController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。